Unify more unit tests
