+++ core/lib/multivm/src/versions/testonly/secp256r1.rs
@@ -3,13 +3,10 @@ use zksync_system_constants::P256VERIFY_PRECOMPILE_ADDRESS;
 use zksync_types::{web3::keccak256, Execute, H256, U256};
 use zksync_utils::h256_to_u256;
 
-use crate::{
-    interface::{ExecutionResult, TxExecutionMode, VmExecutionMode, VmInterface, VmInterfaceExt},
-    vm_fast::tests::tester::VmTesterBuilder,
-};
+use super::{tester::VmTesterBuilder, TestedVm};
+use crate::interface::{ExecutionResult, TxExecutionMode, VmExecutionMode, VmInterfaceExt};
 
-#[test]
-fn test_sekp256r1() {
+pub(crate) fn test_secp256r1<VM: TestedVm>() {
     // In this test, we aim to test whether a simple account interaction (without any fee logic)
     // will work. The account will try to deploy a simple contract from integration tests.
     let mut vm = VmTesterBuilder::new()
@@ -17,7 +14,7 @@ fn test_sekp256r1() {
         .with_execution_mode(TxExecutionMode::VerifyExecute)
         .with_execution_mode(TxExecutionMode::EthCall)
         .with_random_rich_accounts(1)
-        .build();
+        .build::<VM>();
 
     let account = &mut vm.rich_accounts[0];
