+++ core/lib/multivm/src/versions/vm_fast/tests/secp256r1.rs
@@ -1,0 +1,6 @@
+use crate::{versions::testonly::secp256r1::test_secp256r1, vm_fast::Vm};
+
+#[test]
+fn secp256r1() {
+    test_secp256r1::<Vm<_>>();
+}
