+++ core/lib/multivm/src/versions/vm_latest/tests/mod.rs
@@ -4,15 +4,20 @@ use zk_evm_1_5_0::{
     aux_structures::{MemoryPage, Timestamp},
     zkevm_opcode_defs::{ContractCodeSha256Format, VersionedHashLen32},
 };
-use zksync_types::{writes::StateDiffRecord, StorageKey, H256, U256};
-use zksync_utils::{bytecode::hash_bytecode, bytes_to_be_words, h256_to_u256, u256_to_h256};
-use zksync_vm_interface::{CurrentExecutionState, VmExecutionMode, VmExecutionResultAndLogs};
+use zksync_types::{writes::StateDiffRecord, StorageKey, Transaction, H256, U256};
+use zksync_utils::{bytecode::hash_bytecode, bytes_to_be_words, h256_to_u256};
+use zksync_vm_interface::{
+    CurrentExecutionState, L2BlockEnv, VmExecutionMode, VmExecutionResultAndLogs,
+};
 
 use super::{HistoryEnabled, Vm};
 use crate::{
     interface::storage::{InMemoryStorage, StorageView},
     versions::testonly::TestedVm,
-    vm_latest::{constants::BOOTLOADER_HEAP_PAGE, tracers::PubdataTracer, TracerDispatcher},
+    vm_latest::{
+        constants::BOOTLOADER_HEAP_PAGE, tracers::PubdataTracer, types::internals::TransactionData,
+        TracerDispatcher,
+    },
 };
 
 mod bootloader;
@@ -38,7 +43,7 @@ mod prestate_tracer;
 mod refunds;
 mod require_eip712;
 mod rollbacks;
-mod sekp256r1;
+mod secp256r1;
 mod simple_execution;
 mod storage;
 mod tester;
@@ -115,7 +120,7 @@ impl TestedVm for Vm<StorageView<InMemoryStorage>, HistoryEnabled> {
         query.is_fresh
     }
 
-    fn verify_required_bootloader_memory(&self, cells: &[(u32, U256)]) {
+    fn verify_required_bootloader_heap(&self, cells: &[(u32, U256)]) {
         for &(slot, required_value) in cells {
             let current_value = self
                 .state
@@ -126,14 +131,28 @@ impl TestedVm for Vm<StorageView<InMemoryStorage>, HistoryEnabled> {
         }
     }
 
-    fn verify_required_storage(&mut self, cells: &[(StorageKey, H256)]) {
-        for &(key, required_value) in cells {
-            let current_value = self.state.storage.storage.read_from_storage(&key);
-            assert_eq!(
-                u256_to_h256(current_value),
-                required_value,
-                "Invalid value at key {key:?}"
-            );
-        }
+    fn write_to_bootloader_heap(&mut self, cells: &[(usize, U256)]) {
+        let timestamp = Timestamp(self.state.local_state.timestamp);
+        self.state
+            .memory
+            .populate_page(BOOTLOADER_HEAP_PAGE as usize, cells.to_vec(), timestamp)
+    }
+
+    fn read_storage(&mut self, key: StorageKey) -> U256 {
+        self.state.storage.storage.read_from_storage(&key)
+    }
+
+    fn last_l2_block_hash(&self) -> H256 {
+        self.bootloader_state.last_l2_block().get_hash()
+    }
+
+    fn push_l2_block_unchecked(&mut self, block: L2BlockEnv) {
+        self.bootloader_state.push_l2_block(block);
+    }
+
+    fn push_transaction_with_refund(&mut self, tx: Transaction, refund: u64) {
+        let tx = TransactionData::new(tx, false);
+        let overhead = tx.overhead_gas();
+        self.push_raw_transaction(tx, overhead, refund, true)
     }
 }
