+++ core/lib/multivm/src/versions/vm_latest/tests/tester/mod.rs
@@ -1,7 +1,5 @@
 pub(crate) use transaction_test_info::{ExpectedError, TransactionTestInfo, TxModifier};
-pub(crate) use vm_tester::{
-    default_l1_batch, get_empty_storage, InMemoryStorageView, VmTester, VmTesterBuilder,
-};
+pub(crate) use vm_tester::{get_empty_storage, InMemoryStorageView, VmTester, VmTesterBuilder};
 pub(crate) use zksync_test_account::{Account, DeployContractsTx, TxType};
 
 mod inner_state;
